backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-socketio==5.11.2
python-multipart==0.0.9
pydantic==2.6.3
pywebview==4.4.1
pyinstaller==6.5.0
aiofiles==23.2.1
